pyproject.toml
[project]
name = "vldc-bot"
version = "0.12.0"
description = "VLDC Nyan bot"
readme = "README.md"
requires-python = "==3.14.2"
dependencies = [
  "python-telegram-bot[job-queue]==22.6",
  "pymongo==4.16.0",
  "google-cloud-translate==3.24.0",
  "google-cloud-speech==2.36.0",
  "sentry-sdk==2.51.0",
  "Pillow==12.1.0",
  "toml==0.10.2",
  "pip-install==1.3.5",
  "aiohttp==3.13.3",
  "openai==2.16.0",
  "google-genai==1.61.0",
  "googleapis-common-protos==1.72.0",
  "protobuf==6.33.5",
]

[dependency-groups]
dev = [
  "black==26.1.0",
  "flake8==7.3.0",
  "mypy==1.19.1",
  "pylint==4.0.4",
  "pytest==9.0.2",
  "autopep8==2.3.2",
  "debugpy==1.8.20",
]
